bug fix
sometimes PEER ground motion files contain numbers that start with a
period.
e.g. .234 .455 .322
in this case the `delimiter=' '` option that I had specified causes
problems because numpy assumes that the data are separated by a single
space.
My option was actually redundant since passing no option for the
delimiter results in numpy treating any sequence of spaces as a
delimiter.
Removing the option makes the code run in both cases (numbers that do
and don't start with a period)